+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Add support for 0-RTT.
 - Remove `Session.Close()`. Applications need to pass an application error code to the transport using `Session.CloseWithError()`.
+- Make the TLS Cipher Suites configurable (via `tls.Config.CipherSuites`).
 
 ## v0.14.0 (2019-12-04)